Search: authorize search and trash as reads in the chain
The search endpoint is a POST, so Kubernetes parses it as a create on an object
named after the last path segment. Dashboards restated that as a list inside
their own authorizer, which meant any other kind mounting the endpoint would
demand create permission to search it, denying readers and granting search only
to those who could also create.

The restatement moves into the chain, so a kind gets it by being served rather
than by remembering to ask. It happens before the chain rather than inside one
link because more than one link reads the verb: the org role authorizer allows a
viewer to list but not to create.

Trash is covered too. It is a POST at the same shape, so it has the same problem,
and one flag is expected to govern both endpoints.

The path segments move to the API types package. The authorization chain needs
them, and importing the handler package there would add a thousand dependencies
to a package that sits under everything.
