+++ pkg/apis/search/v0alpha1/types.go
@@ -18,6 +18,11 @@ const (
 	KindSearchResults = "SearchResults"
 	KindTrashQuery    = "TrashQuery"
 	KindTrashResults  = "TrashResults"
+
+	// Here rather than beside the routes because the authorization chain needs
+	// them and cannot depend on the handler package.
+	SearchPathSegment = "search"
+	TrashPathSegment  = "trash"
 )
 
 // WhereNode is a single node of the where tree. Exactly one field must be set;
