+++ pkg/registry/apis/search/route.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 	"strings"
 
-	"k8s.io/apiserver/pkg/authorization/authorizer"
 	"k8s.io/kube-openapi/pkg/spec3"
 	"k8s.io/kube-openapi/pkg/validation/spec"
 
@@ -18,8 +17,8 @@ const (
 	ConfigKey     = "enable_search_api"
 )
 
-// searchPathSegment is the last segment of the search endpoint path.
-const searchPathSegment = "search"
+// Aliased so the authorization chain and the route cannot drift apart.
+const searchPathSegment = searchv0.SearchPathSegment
 
 // Route is an endpoint to mount, described in terms the caller's apiserver
 // wiring can consume. Deliberately not Grafana's builder.APIRouteHandler: the
@@ -60,40 +59,6 @@ func capitalize(s string) string {
 		return s
 	}
 	return strings.ToUpper(s[:1]) + s[1:]
-}
-
-// IsSearchRequest reports whether attr describes a call to a kind's search
-// endpoint. Kubernetes parses that path as a create on the kind named "search",
-// which a real create never is: creating an object posts to the collection and
-// so carries no name.
-func IsSearchRequest(attr authorizer.Attributes) bool {
-	return attr.IsResourceRequest() &&
-		attr.GetVerb() == "create" &&
-		attr.GetSubresource() == "" &&
-		attr.GetName() == searchPathSegment
-}
-
-// AsReadAttributes restates a search request as the read it performs. Without
-// this, searching would demand permission to create the kind.
-func AsReadAttributes(attr authorizer.Attributes) authorizer.Attributes {
-	fieldSelector, fieldErr := attr.GetFieldSelector()
-	labelSelector, labelErr := attr.GetLabelSelector()
-	return authorizer.AttributesRecord{
-		User:            attr.GetUser(),
-		Verb:            "list",
-		Namespace:       attr.GetNamespace(),
-		APIGroup:        attr.GetAPIGroup(),
-		APIVersion:      attr.GetAPIVersion(),
-		Resource:        attr.GetResource(),
-		Subresource:     attr.GetSubresource(),
-		ResourceRequest: true,
-		Path:            attr.GetPath(),
-
-		FieldSelectorRequirements: fieldSelector,
-		FieldSelectorParsingErr:   fieldErr,
-		LabelSelectorRequirements: labelSelector,
-		LabelSelectorParsingErr:   labelErr,
-	}
 }
 
 func searchRouteSpec(kindName, version string) *spec3.PathProps {
